MBS-9 Refactor `trials`
* Fix issue with displays when in simulation mode
